cleanup einsum handler
