+++ hls4ml/converters/keras_v3/einsum_dense.py
@@ -8,11 +8,12 @@ if typing.TYPE_CHECKING:
     from keras.api import KerasTensor
 
 
-def strip_batch_dim(equation: str):
+def strip_batch_dim(equation: str, einsum_dense: bool = True):
     """Remove the batch dimension from the equation.
 
     Args:
         equation (str): The einsum equation.
+        einsum_dense (bool): Whether the equation is for EinsumDense layer.
 
     Returns:
         str: The einsum equation without the batch dimension.
@@ -20,12 +21,16 @@ def strip_batch_dim(equation: str):
 
     _inps, out = equation.split('->')
     inp0, inp1 = _inps.split(',')
-    if inp0.startswith('...'):
-        assert out.startswith('...'), f'Error in eq: {equation}: Batch dim mismatch for the input and output.'
+    if einsum_dense:
+        if inp0.startswith('...'):
+            assert out.startswith('...'), f'Error in eq: {equation}: Batch dim mismatch for the input and output.'
+        else:
+            assert inp0[0] == out[0], f'Error in eq: {equation}: Batch dim mismatch for the input and output.'
+            assert inp0[0] not in inp1, f'Error in eq: {equation}: Batch dim is used in the kernel.'
+            inp0, out = inp0[1:], out[1:]
     else:
-        assert inp0[0] == out[0], f'Error in eq: {equation}: Batch dim mismatch for the input and output.'
-        assert inp0[0] not in inp1, f'Error in eq: {equation}: Batch dim is used in the kernel.'
-        inp0, out = inp0[1:], out[1:]
+        assert inp0[0] == inp1[0] == out[0], f'Error in eq: {equation}: Batch dim mismatch for the inputs and output.'
+        inp0, inp1, out = inp0[1:], inp1[1:], out[1:]
     return f'{inp0},{inp1}->{out}'
 
 
@@ -52,7 +57,7 @@ class KV3EinsumDenseHandler(KerasV3LayerHandler):
             f'Error when processing {layer.name}: EinsumDense layer requires fully out shapes'
         # fmt: on
 
-        equation = strip_batch_dim(layer.equation)
+        equation = strip_batch_dim(layer.equation, True)
 
         kernel = self.load_weight(layer, 'kernel')
